requirements.txt
python-dotenv~=1.0.1
langchain-core~=0.3.6
langchain-openai~=0.2.1